isolate the specific implementation of the voting
