+++ app/cells/decidim/decidim_awesome/voting/three_flags_counter/show.erb
