+++ app/cells/decidim/decidim_awesome/voting/three_flags_counter/vote_button.erb
