+++ app/cells/decidim/decidim_awesome/voting/three_flags_proposal/show.erb
