+++ app/cells/decidim/decidim_awesome/voting/three_flags_proposal/vote_block_for.erb
