--- app/cells/decidim/proposals/vote_proposal_modal_cell.rb
@@ -1,30 +1,0 @@
-# frozen_string_literal: true
-
-module Decidim
-  module Proposals
-    class VoteProposalModalCell < Decidim::ViewModel
-      include Decidim::ComponentPathHelper
-      include Decidim::Proposals::Engine.routes.url_helpers
-
-      def show
-        render :show
-      end
-
-      def proposal
-        model
-      end
-
-      def modal_id
-        options[:modal_id] || "voteProposalModal"
-      end
-
-      def from_proposals_list
-        options[:from_proposals_list]
-      end
-
-      def current_component
-        proposal.component
-      end
-    end
-  end
-end
