--- app/packs/src/decidim/decidim_awesome/channels/consumer.js
@@ -1,3 +1,0 @@
-import { createConsumer } from "@rails/actioncable"
-
-export default createConsumer()
