+++ app/packs/stylesheets/decidim/decidim_awesome/awesome_application.scss
@@ -2,7 +2,7 @@
 @import "stylesheets/decidim/decidim_awesome/editors/markdown_editor";
 @import "stylesheets/decidim/decidim_awesome/editors/quill_editor";
 @import "stylesheets/decidim/decidim_awesome/forms/autosave";
-@import "stylesheets/decidim/decidim_awesome/vote/vote";
+@import "stylesheets/decidim/decidim_awesome/voting/three_flags";
 
 .decidim_awesome-custom_fields {
   dd {
