+++ app/packs/stylesheets/decidim/decidim_awesome/voting/three_flags.scss
@@ -1,0 +1,95 @@
+.voting-three-flag {
+  $rectangle-width: 7rem;
+  $green-color: #4CAF50;
+  $yellow-color: #FFC107;
+  $red-color: #F44336;
+
+  .vote-title {
+    text-align: center;
+  }
+
+  .votes_counter {
+    letter-spacing: 0.2rem;
+    font-size: 1.5rem;
+  }
+
+  .rectangle {
+    width: $rectangle-width;
+    height: calc($rectangle-width * 2);
+    display: inline-block;
+    margin: 5px;
+    border-radius: 5px;
+    position: relative;
+
+    &.green {
+      background-color: $green-color;
+    }
+
+    &.yellow {
+      background-color: $yellow-color;
+    }
+
+    &.red {
+      background-color: $red-color;
+    }
+  }
+
+  .voting-block {
+    text-align: center;
+    line-height: 40px;
+  }
+
+  .voting-button {
+    background: none;
+    border: none;
+    color: #fff;
+    font-weight: bold;
+  }
+
+  .fully-opaque {
+    opacity: 1;
+  }
+
+  .semi-opaque {
+    opacity: 0.3;
+    pointer-events: none;
+  }
+
+  .vote-block {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    position: relative;
+
+    .vote-count {
+      font-size: 3rem;
+      font-weight: bold;
+    }
+
+    .vote-label {
+      font-size: 1.5rem;
+      margin-bottom: 0;
+
+      &.green {
+        color: $green-color;
+      }
+
+      &.yellow {
+        color: $yellow-color;
+      }
+
+      &.red {
+        color: $red-color;
+      }
+    }
+
+    .icon--thumb-up {
+      position: absolute;
+      bottom: 0;
+      left: 10%;
+      height: 40px;
+      width: 40px;
+    }
+  }
+}
