+++ app/views/decidim/decidim_awesome/voting/three_flags/_show_vote_button.html.erb
@@ -1,1 +1,1 @@
-<%= cell("decidim/proposals/vote_proposal", proposal, from_proposals_list: true) %>
+<%= cell("decidim/decidim_awesome/voting/three_flags_proposal", proposal, from_proposals_list: true) %>
